Fix Vault flag structure
